pedagogia_graph_code_repo/requirements.txt
aiofiles==24.1.0
aioitertools==0.12.0
aiosqlite==0.21.0
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
arize-phoenix==12.9.0
arize-phoenix-client==1.21.0
arize-phoenix-evals==2.5.0
arize-phoenix-otel==0.13.1
attrs==25.4.0
audioop-lts==0.2.2
Authlib==1.6.5
backoff==2.2.1
bcrypt==5.0.0
beartype==0.22.5
beautifulsoup4==4.14.2
brotli==1.2.0
bs4==0.0.2
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.3.0
clang==20.1.5
click==8.3.0
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
cyclopts==4.2.1
ddgs==9.7.1
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.2
durationpy==0.10
email-validator==2.3.0
esprima==4.0.1
exceptiongroup==1.3.0
fastapi==0.121.0
ffmpy==1.0.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
google-auth==2.42.0
googleapis-common-protos==1.71.0
gradio==5.49.1
gradio_client==1.13.3
graphql-core==3.2.7
greenlet==3.2.4
groovy==0.1.2
grpc-interceptor==0.15.4
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
javalang==0.13.0
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.6.0
kiwisolver==1.4.9
kubernetes==34.1.0
langchain-core==1.0.1
langchain-text-splitters==1.0.0
langfuse==3.9.0
langsmith==0.4.38
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mcp==1.10.1
mcpadapt==0.1.20
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.6.1
openapi-pydantic==0.5.1
openinference-instrumentation==0.1.42
openinference-instrumentation-smolagents==0.1.19
openinference-semantic-conventions==0.1.25
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pathable==0.4.4
pathvalidate==3.3.1
pillow==11.3.0
pip==25.3
platformdirs==4.5.0
pluggy==1.6.0
ply==3.11
posthog==5.4.0
primp==0.15.0
prometheus_client==0.23.1
protobuf==6.33.0
psutil==7.1.3
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
pydantic-settings==2.11.0
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pyperclip==1.11.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pystache==0.6.8
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.28.0
rsa==4.9.1
ruff==0.14.5
safehttpx==0.1.7
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
SecretStorage==3.4.0
semantic-version==2.10.0
sentence-transformers==5.1.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smolagents==1.22.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
SQLAlchemy==2.0.44
sqlean.py==3.50.4.5
sse-starlette==3.0.3
starlette==0.49.3
strawberry-graphql==0.270.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
tree-sitter==0.25.2
tree-sitter-rust==0.24.0
triton==3.5.0
typer==0.20.0
typer-slim==0.20.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.38.0
uvloop==0.22.1
validators==0.35.0
watchfiles==1.1.1
weaviate-client==4.17.0
websocket-client==1.9.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
zipp==3.23.0
zstandard==0.25.0
